refactor: extract common helpers
